moved urls from programs to solicitations
