+++ commcare_connect/templates/program/response_review.html
@@ -16,11 +16,11 @@
             <i class="fa-solid fa-chevron-right"></i>
         </li>
         <li>
-            <a href="{% url 'program:solicitation_dashboard' org_slug=request.org.slug pk=program.pk %}">Solicitations</a>
+            <a href="{% url 'org_solicitations:program_dashboard' org_slug=request.org.slug pk=program.pk %}">Solicitations</a>
             <i class="fa-solid fa-chevron-right"></i>
         </li>
         <li>
-            <a href="{% url 'program:response_list' org_slug=request.org.slug pk=program.pk solicitation_pk=solicitation.pk %}">{{ solicitation.title|truncatechars:20 }}</a>
+            <a href="{% url 'org_solicitations:program_response_list' org_slug=request.org.slug pk=program.pk solicitation_pk=solicitation.pk %}">{{ solicitation.title|truncatechars:20 }}</a>
             <i class="fa-solid fa-chevron-right"></i>
         </li>
         <li>
@@ -68,7 +68,7 @@
                 </div>
 
                 <div class="flex items-center gap-x-4">
-                    <a href="{% url 'program:response_list' org_slug=request.org.slug pk=program.pk solicitation_pk=solicitation.pk %}"
+                    <a href="{% url 'org_solicitations:program_response_list' org_slug=request.org.slug pk=program.pk solicitation_pk=solicitation.pk %}"
                        class="button button-md outline-style">
                         <i class="fa-solid fa-arrow-left mr-2"></i>
                         Back to Responses
