+++ commcare_connect/templates/solicitations/solicitation_form.html
@@ -15,7 +15,7 @@
             <i class="fa-solid fa-chevron-right"></i>
         </li>
         <li>
-            <a href="{% url 'program:solicitation_dashboard' org_slug=request.org.slug pk=program.pk %}">Solicitations</a>
+            <a href="{% url 'org_solicitations:program_dashboard' org_slug=request.org.slug pk=program.pk %}">Solicitations</a>
             <i class="fa-solid fa-chevron-right"></i>
         </li>
         <li>
@@ -59,7 +59,7 @@
             </div>
 
             <div class="flex items-center gap-x-4">
-                <a href="{% url 'program:solicitation_dashboard' org_slug=request.org.slug pk=program.pk %}"
+                <a href="{% url 'org_solicitations:program_dashboard' org_slug=request.org.slug pk=program.pk %}"
                    class="button button-md outline-style">
                     <i class="fa-solid fa-arrow-left mr-2"></i>
                     Back to Dashboard
@@ -255,7 +255,7 @@
 
             <!-- Action Buttons -->
             <div class="flex justify-end gap-4 pt-6 border-t border-gray-200">
-                <a href="{% url 'program:solicitation_dashboard' org_slug=request.org.slug pk=program.pk %}"
+                <a href="{% url 'org_solicitations:program_dashboard' org_slug=request.org.slug pk=program.pk %}"
                    class="button button-md outline-style">
                     Cancel
                 </a>
